# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notification": {
      "get": {
        "operationId": "NotificationController_getUserNotifications",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications list"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get paginated notifications for current user",
        "tags": [
          "Notification"
        ]
      }
    },
    "/notification/unread-count": {
      "get": {
        "operationId": "NotificationController_getUnreadCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Unread notification count"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get unread notification count for current user",
        "tags": [
          "Notification"
        ]
      }
    },
    "/notification/read-all": {
      "patch": {
        "operationId": "NotificationController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All marked as read"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mark all notifications as read for current user",
        "tags": [
          "Notification"
        ]
      }
    },
    "/notification/{id}/read": {
      "patch": {
        "operationId": "NotificationController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Notification ID UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mark a single notification as read",
        "tags": [
          "Notification"
        ]
      }
    },
...
    "/order/initialize": {
      "post": {
        "operationId": "OrderController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order initialized with payment link / reference"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Initialize a new checkout order / payment intent",
        "tags": [
          "Order"
        ]
      }
    },
...
    "/order/paystack/verify": {
      "get": {
        "operationId": "OrderController_verifyPayment",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "query",
            "description": "Paystack transaction reference",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment verification result"
          }
        },
        "summary": "Verify a Paystack payment reference",
        "tags": [
          "Order"
        ]
      }
    },
...
    "/order/approved/all/user": {
      "get": {
        "operationId": "OrderController_getLoginUserAllApprovedOrdersPagination",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User approved orders"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get current user approved orders",
        "tags": [
          "Order"
        ]
      }
    },
    "/order/notapproved/all/user": {
      "get": {
        "operationId": "OrderController_getLoginUserAllNotApprovedOrdersPagination",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User pending orders"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get current user pending / non-approved orders",
        "tags": [
          "Order"
        ]
      }
    },
...
    "/order/pag/all/user": {
      "get": {
        "operationId": "OrderController_getAllLoginUserOrderPagination",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User paginated orders"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all orders of currently logged in user",
        "tags": [
          "Order"
        ]
      }
    },
    "/order/pag/all/{userId}": {
      "get": {
        "operationId": "OrderController_findAllUserOrderPagination",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User orders"
          }
        },
        "summary": "Get all orders of a specific user",
        "tags": [
          "Order"
        ]
      }
    },
...
    "/order/{id}": {
      "get": {
        "operationId": "OrderController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details"
          }
        },
        "summary": "Get order by ID",
        "tags": [
          "Order"
        ]
      },
      "patch": {
        "operationId": "OrderController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated"
          }
        },
        "summary": "Update order status / details",
        "tags": [
          "Order"
        ]
      },
      "delete": {
        "operationId": "OrderController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted"
          }
        },
        "summary": "Delete order by ID",
        "tags": [
          "Order"
        ]
      }
    },
    "/order/{id}/comment/new": {
      "patch": {
        "operationId": "OrderController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment added"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Add a new comment to an order",
        "tags": [
          "Order"
        ]
      }
    },
    "/order/{id}/comment": {
      "patch": {
        "operationId": "OrderController_addCommentFormData",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileUrl": {
                    "type": "string",
                    "format": "binary",
                    "description": "Attachment file (max 50MB)"
                  },
                  "comment": {
                    "type": "string",
                    "description": "Comment text",
                    "example": "Please check this revision"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment with attachment added"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Add comment with file attachment to order",
        "tags": [
          "Order"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mejarc API",
    "description": "Comprehensive REST API documentation for Mejarc Backend services, including Authentication, User Management, Agent Management, Marketplace, Custom Design Wizard, Orders, Payments, Wallet & Withdrawals, Notifications, and Real-Time Chat.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter your Bearer token in format: Bearer <token>",
        "in": "header"
      }
    },
    "schemas": {
...
      "OrderItemDTO": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "totalQuantity": {
            "type": "number"
          },
          "totalPrice": {
            "type": "number"
          }
        },
        "required": [
          "productId",
          "totalQuantity",
          "totalPrice"
        ]
      },
      "BillingInfoDTO": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          }
        }
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "redirect_url": {
            "type": "string"
          },
          "projectDsc": {
            "type": "string"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDTO"
            }
          },
          "amountPaid": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "billingInfo": {
            "$ref": "#/components/schemas/BillingInfoDTO"
          }
        },
        "required": [
          "redirect_url",
          "projectDsc",
          "orderItems",
          "billingInfo"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "redirect_url": {
            "type": "string"
          },
          "projectDsc": {
            "type": "string"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDTO"
            }
          },
          "amountPaid": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "billingInfo": {
            "$ref": "#/components/schemas/BillingInfoDTO"
          }
        }
      },
      "AddCommentDto": {
        "type": "object",
        "properties": {
          "commentText": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          }
        },
        "required": [
          "commentText",
          "fileUrl"
        ]
      },
...
    }
  }
}